debug_viewer/requirements.txt
# Debug viewer dependencies

# Import the main library
# Install with: pip install -e ../procedural_building

# 3D visualization with pygame + OpenGL
pygame>=2.5.0
PyOpenGL>=3.1.6
PyOpenGL-accelerate>=3.1.6  # Optional but recommended for performance
